I have added ownership to the texts we receive ensuring we only get from the actual owner

+++ dbhelper.py
@@ -7,23 +7,25 @@ class DBHelper:
         self.conn=sqlite3.connect(dbname)
 #this method creates a table called items with column called description
     def setUp(self):
-        stmt="CREATE TABLE IF NOT EXISTS items (description text)"
+        print("creating table")
+        stmt="CREATE TABLE IF NOT EXISTS items (description text,owner text)"
         self.conn.execute(stmt)
         self.conn.commit()
 
 #method to add to the database
-    def add_item(self,item_text):
-        stmt="INSERT INTO items (description) VALUES (?)"
-        args=(item_text, )
+    def add_item(self,item_text,owner):
+        stmt="INSERT INTO items (description,owner) VALUES (?,?)"
+        args=(item_text,owner)
         self.conn.execute(stmt,args)
         self.conn.commit()
 #method to delete from the database
-    def delete_item(self,item_text):
-        stmt="DELETE FROM items WHERE description =(?)"
-        args=(item_text, )
+    def delete_item(self,item_text,owner):
+        stmt="DELETE FROM items WHERE description =(?) and owner = (?)"
+        args=(item_text,owner)
         self.conn.execute(stmt,args)
         self.conn.commit()
 #method to return list of items from the db
-    def get_items(self):
-        stmt="SELECT description FROM items"
-        return [x[0] for x in self.conn.execute(stmt)]
+    def get_items(self,owner):
+        stmt="SELECT description FROM items WHERE owner = (?)"
+        arg=(owner,)
+        return [x[0] for x in self.conn.execute(stmt,arg)]
